projects/minimalistic/requirements.txt
django
django-des